abstract some helpers
